version 1.4

+++ fct_gestion_app/js/usuarios_llamadasAPI.js
@@ -1,3 +1,10 @@
+// Mensajes
+const mensajeVacio = "Los campos no pueden estar vacíos."
+const mensajeCorreo = "El correo electrónico debe ser válido."
+const mensajeContrasena = "La contraseña debe tener al menos una longitud de 8 caracteres."
+const mensajePassRepetida = "La contraseñas son distintas."
+const mensajeDNITelefono = "Tanto el campo de DNI como el campo de Teléfono deben tener 9 caracteres."
+
 // Función para pintar usuarios en forma de tabla. Tipo de usuario es el rol
 async function getUsuarios(tipoUsuario) {
     let div = document.getElementById('usuarios')
@@ -5,7 +12,7 @@ async function getUsuarios(tipoUsuario) {
     // creamos la tabla
     let tabla = document.createElement('table')
     tabla.classList.add('table')
-
+    
     let encabezadosTabla = ['Nombre', 'Apellidos', 'Fecha nacimiento', 'DNI', 'Correo', 'Teléfono', 'Editar', 'Borrar']
     let encabezadosFilas = document.createElement('tr')
 
@@ -32,8 +39,8 @@ async function getUsuarios(tipoUsuario) {
                     <td>${usuario.dni}</td>
                     <td>${usuario.email}</td>
                     <td>${usuario.telefono}</td>
-                    <td><input type='submit' onclick='' value='✏'></td>
-                    <td><input type='submit' onclick='deleteUsuario(${usuario.id})' value='❌'></td>`
+                    <td><input type='button' onclick='' value='✏' class='botonEditar'></td>
+                    <td><input type='button' onclick='deleteUsuario(${usuario.id})' value='❌' class='botonBorrar'></td>`
                 tabla.appendChild(fila)
             }
         })
@@ -46,7 +53,7 @@ async function getUsuarios(tipoUsuario) {
 
 // Función para registrar un usuario
 async function registerUsuario() {
-    const form = document.getElementById('formularioRegister')
+    const form = document.getElementById('formularioRegistro')
   
     form.addEventListener('submit', async (event) => {
         event.preventDefault() // prevenir que el formulario se envíe por defecto
@@ -54,54 +61,76 @@ async function registerUsuario() {
         // cogemos el token del profesor
         let token = localStorage.getItem('token')
 
-        // si los datos son correctos
-        if(validarRegistro() == true) {
-            let miHeaders = new Headers()
-            miHeaders.append(`Authorization`, `Bearer ${token}`)
-        
-            // hacemos el registro
-            let datos = JSON.stringify({
-                "nombre": nombre,
-                "apellidos": apellidos,
-                "fecha_nacimiento": fecha_nacimiento,
-                "dni": dni,
-                "email": email,
-                "telefono": telefono,
-                "password": password2,
-                "rol_id": 2
-            });
-        
-            let requestRegistro = {
-                method: 'POST',
-                headers: miHeaders,
-                body: datos,
-                redirect: 'follow'
-            };
-        
-            fetch(`${API_BASE_URL}/register`, requestRegistro)
-                .then(response => response.text())
-                .then(result => console.log(result))
-                .catch(error => console.log('error', error))
-        }
+        let miHeaders = new Headers()
+        miHeaders.append("Content-Type", "application/json")
+        miHeaders.append(`Authorization`, `Bearer ${token}`)
+
+        // hacemos el registro
+        let datos = JSON.stringify({
+            "nombre": `${document.getElementById('nombre').value}`,
+            "apellidos": `${document.getElementById('apellidos').value}`,
+            "fecha_nacimiento": `${document.getElementById('fecha_nacimiento').value}`,
+            "dni": `${document.getElementById('dni').value}`,
+            "email": `${document.getElementById('correo').value}`,
+            "telefono": `${document.getElementById('telefono').value}`,
+            "password": `${document.getElementById('password2').value}`,
+            "rol_id": "2"
+        });
+
+        let requestRegistro = {
+            method: 'POST',
+            headers: miHeaders,
+            body: datos,
+            redirect: 'follow'
+        };
+
+        fetch(`${API_BASE_URL}/register`, requestRegistro)
+            .then(response => {
+                if(response.ok) {
+                    alert("Alumno creado correctamente")
+                    window.location.href = "http://127.0.0.1:3000/fct_gestion_app/gestion_alumnos.html"
+                }
+                response.text()
+            })
+            .then(result => console.log(result))
+            .catch(error => console.log('error', error))
     })
 }
 
 // Función para validar los campos de registros
 function validarRegistro() {
-    if (nombre == "" || apellidos == "" || fecha_nacimiento == "" || dni == "" || 
-    email == "" || telefono == "" || password1 == "" || password2 == "") {
-        alert("Todos los campos son obligatorios")
-        return false
+    limpiarOutput("errores")
+    let campoErrores = document.getElementById("errores")
+    let validar = true
+
+    if (!validarCamposVacios(document.getElementById('nombre').value) || !validarCamposVacios(document.getElementById('apellidos').value) || 
+    !validarCamposVacios(document.getElementById('password1').value) || !validarCamposVacios(document.getElementById('password2').value) || 
+    !validarCamposVacios(document.getElementById('fecha_nacimiento').value) || !validarCamposVacios(document.getElementById('dni').value) || 
+    !validarCamposVacios(document.getElementById('telefono').value) || !validarCamposVacios(document.getElementById('correo').value)) {
+        campoErrores.innerHTML += `${mensajeVacio} <br>`
+        validar = false
     }
 
-    // Verificar si las contraseñas coinciden
-    if (password1 !== password2) {
-        alert("Las contraseñas no coinciden")
-        return false
+    if (!validarCorreo(document.getElementById('correo').value)) {
+        campoErrores.innerHTML += `${mensajeCorreo} <br>`
+        validar = false
     }
 
-    // Si todo está correcto, enviar el formulario
-    return true
+    if (!validarPassword(document.getElementById('password1').value)) {
+        campoErrores.innerHTML += `${mensajeContrasena} <br>`
+        validar = false
+    }
+
+    if (!validar2Password(document.getElementById('password1').value, document.getElementById('password2').value)) {
+        campoErrores.innerHTML += `${mensajePassRepetida} <br>`
+        validar = false
+    }
+
+    if (!validarTamanio(document.getElementById('dni').value, 9, 9) ||
+    !validarTamanio(document.getElementById('telefono').value, 9, 9)) {
+        campoErrores.innerHTML += `${mensajeDNITelefono} <br>`
+        validar = false
+    }
 }
 
 // Función para editar un usuario
@@ -111,22 +140,28 @@ async function putUsuario(id) {
   
 // Función para eliminar un usuario
 async function deleteUsuario(id) {
-    mensajeConfirmacion("usuario")
-    let token = localStorage.getItem('token')
+    let mensajeConfirmacion = confirm("¿Está seguro que desea eliminar a este usuario?")
+    if(mensajeConfirmacion) {
+        let token = localStorage.getItem('token')
 
-    let miHeaders = new Headers()
-    miHeaders.append(`Authorization`, `Bearer ${token}`)
-
-    let requestOptions = {
-        method: 'DELETE',
-        headers: miHeaders,
-        redirect: 'follow'
-    };
-
-    fetch(`${API_BASE_URL}/usuarios/${id}`, requestOptions)
-        .then(response => response.text())
-        .then(result => console.log(result))
-        .catch(error => console.log('error', error))
+        let miHeaders = new Headers()
+        miHeaders.append(`Authorization`, `Bearer ${token}`)
+    
+        let requestOptions = {
+            method: 'DELETE',
+            headers: miHeaders,
+            redirect: 'follow'
+        };
+    
+        fetch(`${API_BASE_URL}/usuarios/${id}`, requestOptions)
+            .then(response => response.text())
+            .then(result => {
+                console.log(result)
+                alert("Alumno borrado correctamente")
+                window.location.href = "http://127.0.0.1:3000/fct_gestion_app/gestion_alumnos.html"
+            })
+            .catch(error => console.log('error', error))
+    }
 }
 
 // Función para cerrar sesión
@@ -138,7 +173,7 @@ async function logout() {
     
     // lo pasamos en formato JSON
     let datos = JSON.stringify({
-        "token": token
+        "token": `${token}`
     });
     
     let requestOptions = {
@@ -160,8 +195,7 @@ async function consultarToken() {
     let token = localStorage.getItem('token')
     // si no existe, redirigimos al inicio de sesión
     if(token == null || token == "") {
-        console.log(mensajeInicioSesion)
-        window.location.href = "http://127.0.0.1:3000/fct_gestion_app/login.html";
+        window.location.href = "http://127.0.0.1:3000/fct_gestion_app/login.html"
     }
 
     // Configurar los encabezados para la solicitud user POST
@@ -189,9 +223,13 @@ async function consultarToken() {
         .then(result => {
             let resultadoUsuario = result.usuario
             document.getElementById("nombre-usuario").textContent = `Bienvenido/a ${resultadoUsuario.nombre} 👋`
-            nombre.innerHTML = resultadoUsuario.nombre
-            apellidos.innerHTML = resultadoUsuario.apellidos
 
+            document.getElementById('nombre').textContent = resultadoUsuario.nombre
+            document.getElementById('apellidos').textContent = resultadoUsuario.apellidos
+            document.getElementById('fecha_nacimiento').textContent = resultadoUsuario.fecha_nacimiento
+            document.getElementById('dni').textContent = resultadoUsuario.dni
+            document.getElementById('email').textContent = resultadoUsuario.email
+            document.getElementById('telefono').textContent = resultadoUsuario.telefono
             return result
         })
         .catch(error => console.log('Error al obtener usuario', error))
